Module Unicorn_jsoo.E

HTML events.

Just like attributes, events are specified inside the HTML node children and attach to their closest parent.

val click : ( 'a -> 'a ) -> 'a t

click fn invokes the callback fn when its HTML parent node is clicked.

∀ f g. click f & click g == click (fun x -> g (f x))

∀ i f. iso i (click f) == click (Iso.map i f) and same for lenses and prisms

Unsafe

Consider submitting a PR if you find yourself using this!

val make : _ Js_of_ocaml.Dom.event Js_of_ocaml.Js.t as 'e Js_of_ocaml.Dom_html.Event.typ -> ( 'e -> 'a -> 'a ) -> 'a t

make ev callback attaches a new event handler to the HTML parent node.